refactor: add model and repository computer
