+++ .gitignore
@@ -1,4 +1,5 @@
 bin
 obj
 .vscode
+database.db
 *.db
